Converted altimeter
Made altitude an int32_t and initial implementation of a simple state machine.

+++ modroc_controller/include/altimeter.h
@@ -28,4 +28,4 @@ Error_Returns altimeter_reset();
 
 Error_Returns altimeter_update_altitude();
 
-double altimeter_get_delta();
+int32_t altimeter_get_delta();
